Remove all HashCalculator updating from NewGameBoard, attaching and running callbacks
